+++ praw/models/comment_forest.py
@@ -21,7 +21,7 @@ class CommentForest(object):
             if isinstance(comment, MoreComments):
                 heappush(more_comments, comment)
                 if parent:
-                    parent.replies.remove(comment)
+                    parent.replies._comments.remove(comment)
                 else:
                     tree.remove(comment)
             else:
@@ -33,14 +33,16 @@ class CommentForest(object):
         """Return the comment at position ``index`` in the list."""
         return self._comments[index]
 
-    def __init__(self, submission):
+    def __init__(self, submission, comments=None):
         """Initialize a CommentForest instance.
 
         :param submission: An instance of :class:`~.Subreddit` that is the
             parent of the comments.
+        :param comments: Initialize the Forest with a list of comments
+            (default: None).
+
         """
-        self._comments = None
-        self._comments_by_id = {}
+        self._comments = comments
         self._submission = submission
 
     def __len__(self):
@@ -48,13 +50,14 @@ class CommentForest(object):
         return len(self._comments)
 
     def _insert_comment(self, comment):
-        assert comment.name not in self._comments_by_id
+        assert comment.name not in self._submission._comments_by_id
         comment.submission = self._submission
         if comment.is_root:
             self._comments.append(comment)
         else:
-            assert comment.parent_id in self._comments_by_id
-            self._comments_by_id[comment.parent_id].replies.append(comment)
+            assert comment.parent_id in self._submission._comments_by_id
+            parent = self._submission._comments_by_id[comment.parent_id]
+            parent.replies._comments.append(comment)
 
     def _update(self, comments):
         self._comments = comments
@@ -80,16 +83,35 @@ class CommentForest(object):
     def replace_more(self, limit=32, threshold=0):
         """Update the comment forest by resolving instances of MoreComments.
 
-        :param limit: The maximum number of MoreComments instances to
+        :param limit: The maximum number of :class:`.MoreComments` instances to
             replace. Each replacement requires 1 API request. Set to ``None``
-            to have no limit, or to ``0`` to remove all MoreComments instances
-            without additional requests (Default: 32).
+            to have no limit, or to ``0`` to remove all :class:`.MoreComments`
+            instances without additional requests (Default: 32).
         :param threshold: The minimum number of children comments a
-            MoreComments instance must have in order to be
-            replaced. MoreComments instances that represent "continue this
-            thread" links unfortunately appear to have 0 children. (Default:
-            0).
-        :returns: A list of MoreComments instances that were not replaced.
+            :class:`.MoreComments` instance must have in order to be
+            replaced. :class:`.MoreComments` instances that represent "continue
+            this thread" links unfortunately appear to have 0
+            children. (Default: 0).
+
+        :returns: A list of :class:`.MoreComments` instances that were not
+            replaced.
+
+        For example, to replace up to 32 :class:`.MoreComments` instances of a
+        submission try:
+
+        .. code:: python
+
+           submission = reddit.submission('3hahrw')
+           submission.comments.replace_more()
+
+        Alternatively, to replace :class:`.MoreComments` instances within the
+        replies of a single comment try:
+
+        .. code:: python
+
+           comment = reddit.comment('d8r4im1')
+           comment.refresh()
+           comment.replace_more()
 
         """
         remaining = limit